fix: rename marketDetailPage to ethrisePage
