+++ pages/markets/ethrise/index.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from "next";
 
 // Import components
 import ETHRISEPage from "../../../components/v1/MarketDetailPage/ETHRISEPage";
-import EthrisePageContainer from "../../../modules/marketDetailPage/EthrisePageContainer";
+import EthrisePageContainer from "../../../modules/ethrisePage/EthrisePageContainer";
 
 const ETHRISE: NextPage = () => {
     return <EthrisePageContainer />;
